- refactor: styled.js => detail.styled.js 로 이름 변경
