+++ src/pages/detail/detail.styled.js
